Refactored app's directory structure
