+++ app/src/main/java/com/chrsrck/quakemap/ui/EarthquakeListFragment.kt
@@ -1,4 +1,4 @@
-package com.chrsrck.quakemap
+package com.chrsrck.quakemap.ui
 
 import android.content.Context
 import android.os.Bundle
@@ -9,7 +9,7 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
+import com.chrsrck.quakemap.R
 
 import com.chrsrck.quakemap.dummy.DummyContent
 import com.chrsrck.quakemap.dummy.DummyContent.DummyItem
