+++ app/src/main/res/layout/fragment_earthquake_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="android.support.v7.widget.LinearLayoutManager"
-    tools:context=".EarthquakeListFragment"
+    tools:context=".ui.EarthquakeListFragment"
     tools:listitem="@layout/fragment_earthquake" />
